MNT #148 bluesky8IDI now starts, no errors
