+++ profile_bluesky/startup/instrument/plans/__init__.py
@@ -3,7 +3,7 @@
 local, custom Bluesky plans (scans) and other functions
 """
 
-from .alignment import *
+from .alignments import *
 from .flux_calculations import *
 from .lineup_tweak import * 
 from .move_diodes import *
