+++ profile_bluesky/startup/instrument/plans/functions.py
@@ -13,7 +13,7 @@ def taylor_series(x, coefficients):
     """
     compute a Taylor series expansion at x
     
-    a0 + x*(a1 + x*(a2+x*0)
+    f(x, a) = a0 + x*(a1 + x*(a2+x*0)
     """
     v = 0
     for a in reversed(coefficients):
